display inventory list, AB#1157
