+++ src/main/java/client/StrainXMain.java
@@ -1,13 +1,14 @@
 package client;
 
-import ui.UIEnterGame;
-
+import ui.gui.GUIEntry;
 import java.io.IOException;
 
 public class StrainXMain {
 
     public static void main(String[] args) throws IOException, InterruptedException {
 
-        UIEnterGame.displayEnterGame();
+//        UIEnterGame.displayEnterGame();
+        GUIEntry.start();
+
     }
 }
